make main menu use the new menu system
