+++ main_d1/newmenu.cpp
@@ -35,6 +35,7 @@ COPYRIGHT 1993-1998 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 #include "game.h"
 #include "stringtable.h"
 #include "newmenu.h"
+#include "newmenu_windows.h"
 #include "gamefont.h"
 #include "network.h"
 #include "iff/iff.h"
@@ -70,17 +71,9 @@ COPYRIGHT 1993-1998 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 
 int Newmenu_first_time = 1;
 
-struct bkg
-{
-	grs_canvas* menu_canvas;
-	grs_bitmap* saved;			// The background under the menu.
-	grs_bitmap* background;
-};
-
 grs_canvas* nm_canvas;
 
 constexpr int MESSAGEBOX_TEXT_SIZE = 300;	// How many characters in messagebox
-constexpr int MAX_TEXT_WIDTH = 200;			// How many pixels wide a input box can be
 
 grs_bitmap nm_background;
 
@@ -94,10 +87,7 @@ std::stack <std::unique_ptr<nm_window>> nm_open_windows;
 
 grs_canvas* nm_get_top_canvas()
 {
-	if (menu_canvas_stack.empty())
-		return nullptr;
-
-	return menu_canvas_stack.top();
+	return nm_canvas;
 }
 
 void newmenu_close(void) 
@@ -352,91 +342,6 @@ void nm_string_inputbox(bkg* b, int w, int x, int y, char* text, int current)
 	}
 }
 
-void draw_item(bkg* b, newmenu_item* item, int is_current, bool tiny)
-{
-	if (tiny)
-	{
-		if (is_current)
-			gr_set_fontcolor(gr_find_closest_color_current(57, 49, 20), -1);
-		else
-			gr_set_fontcolor(gr_find_closest_color_current(29, 29, 47), -1);
-
-		//Not quite sure why Descent 2 does this
-		if (item->text[0] == '\t')
-			gr_set_fontcolor(gr_find_closest_color_current(63, 63, 63), -1);
-	}
-	else
-	{
-		if (is_current)
-			grd_curcanv->cv_font = CURRENT_FONT;
-		else if (item->type == NM_TYPE_TEXT)
-			grd_curcanv->cv_font = TEXT_FONT; //Need to do this here for tiny rendering to work right
-		else
-			grd_curcanv->cv_font = NORMAL_FONT;
-	}
-
-	switch (item->type) 
-	{
-	case NM_TYPE_TEXT:
-	case NM_TYPE_MENU:
-		nm_string(b, item->w, item->x, item->y, item->text);
-		break;
-	case NM_TYPE_SLIDER: 
-	{
-		if (item->value < item->min_value) item->value = item->min_value;
-		if (item->value > item->max_value) item->value = item->max_value;
-		sprintf(item->saved_text, "%s\t%s", item->text, SLIDER_LEFT);
-		for (int j = 0; j < (item->max_value - item->min_value + 1); j++) 
-		{
-			strcat(item->saved_text, SLIDER_MIDDLE);
-		}
-		strcat(item->saved_text, SLIDER_RIGHT);
-
-		item->saved_text[item->value + 1 + strlen(item->text) + 1] = SLIDER_MARKER[0];
-
-		nm_string_slider(b, item->w, item->x, item->y, item->saved_text);
-	}
-						 break;
-	case NM_TYPE_INPUT_MENU:
-		if (item->group == 0) 
-		{
-			nm_string(b, item->w, item->x, item->y, item->text);
-		}
-		else 
-		{
-			nm_string_inputbox(b, item->w, item->x, item->y, item->text, is_current);
-		}
-		break;
-	case NM_TYPE_INPUT:
-		nm_string_inputbox(b, item->w, item->x, item->y, item->text, is_current);
-		break;
-	case NM_TYPE_CHECK:
-		nm_string(b, item->w, item->x, item->y, item->text);
-		if (item->value)
-			nm_rstring(b, item->right_offset, item->x, item->y, CHECKED_CHECK_BOX);
-		else
-			nm_rstring(b, item->right_offset, item->x, item->y, NORMAL_CHECK_BOX);
-		break;
-	case NM_TYPE_RADIO:
-		nm_string(b, item->w, item->x, item->y, item->text);
-		if (item->value)
-			nm_rstring(b, item->right_offset, item->x, item->y, CHECKED_RADIO_BOX);
-		else
-			nm_rstring(b, item->right_offset, item->x, item->y, NORMAL_RADIO_BOX);
-		break;
-	case NM_TYPE_NUMBER: 
-	{
-		char text[10];
-		if (item->value < item->min_value) item->value = item->min_value;
-		if (item->value > item->max_value) item->value = item->max_value;
-		nm_string(b, item->w, item->x, item->y, item->text);
-		sprintf(text, "%d", item->value);
-		nm_rstring(b, item->right_offset, item->x, item->y, text);
-	}
-		break;
-	}
-
-}
 
 char* Newmenu_allowed_chars = NULL;
 
@@ -488,636 +393,28 @@ int newmenu_do2(const char* title, const char* subtitle, int nitems, newmenu_ite
 	return newmenu_do3(title, subtitle, nitems, item, subfunction, citem, filename, -1, -1);
 }
 
-//-----------------------------------------------------------------------------
-// nm_menu
-//
-// Implements a normal menu that can have a couple of different widgets
-//-----------------------------------------------------------------------------
 
-class nm_menu : public nm_window
+void newmenu_open(const char* title, const char* subtitle, std::vector<newmenu_item>& items, void (*subfunction)(int nitems, newmenu_item* items, int* last_key, int citem), bool (*choicefunc)(int choice, newmenu_item* item), int citem, const char* filename, int width, int height, bool tiny_mode)
 {
-	std::string title, subtitle, bg_filename;
-	std::vector<newmenu_item> items;
-	int choice, old_choice;
-	int tw, th;
-	int w, h, x, y;
-	bool tiny, all_text;
-	bkg bg;
+	nm_open_windows.push(std::make_unique<nm_menu>(items, title, subtitle, subfunction, choicefunc, citem, filename, width, height, tiny_mode));
+}
 
-	//Callback executed when a choice is selected.
-	//Returns true if the menu should remain open, or false if it should close. 
-	bool (*choice_callback)(int choice);
-	//Callback executed each frame, to update and redraw contents. 
-	void (*newmenu_callback)(int nitems, newmenu_item* items, int* last_key, int citem);
-public:
-	nm_menu(std::vector<newmenu_item>& source_items, const char* new_title, const char* new_subtitle,
-		void (*subfunction)(int nitems, newmenu_item* items, int* last_key, int citem), bool (*choicefunc)(int choice),
-		int citem, const char* filename, int width, int height, bool tiny_mode)
-	{
-		for (newmenu_item& item : source_items)
-			items.push_back(item);
-
-		grs_canvas* save_canvas = grd_curcanv;
-		gr_set_current_canvas(nm_canvas);
-
-		choice = old_choice = 0;
-
-		Assert(choicefunc != nullptr);
-
-		bg = {};
-		all_text = false;
-		choice_callback = choicefunc;
-		newmenu_callback = subfunction;
-
-		if (new_title)
-			title = std::string(new_title);
-		if (new_subtitle)
-			subtitle = std::string(new_subtitle);
-
-		//Temp variables
-		int string_width, string_height, average_width;
-
-		tw = th = 0;
-
-		//Find the size of the title and subtitle.
-		if (title.size())
-		{
-			grd_curcanv->cv_font = TITLE_FONT;
-			gr_get_string_size(title.c_str(), &string_width, &string_height, &average_width);
-			tw = string_width;
-			th = string_height;
-		}
-		if (subtitle.size())
-		{
-			grd_curcanv->cv_font = SUBTITLE_FONT;
-			gr_get_string_size(subtitle.c_str(), &string_width, &string_height, &average_width);
-			if (string_width > tw)
-				tw = string_width;
-			th += string_height;
-		}
-
-		th += 8;		//put some space between titles & body
-
-		tiny = tiny_mode;
-		if (tiny)
-			grd_curcanv->cv_font = SMALL_FONT;
-		else
-			grd_curcanv->cv_font = NORMAL_FONT;
-
-		w = 0; h = th;
-		int aw = 0;
-
-		//Find ths size of all the menu items
-		for (newmenu_item& item : items)
-		{
-			item.redraw = 1;
-			item.y = h;
-			gr_get_string_size(item.text, &string_width, &string_height, &average_width);
-			item.right_offset = 0;
-
-			item.saved_text[0] = '\0';
-
-			if (item.type == NM_TYPE_SLIDER)
-			{
-				int w1, h1, aw1;
-				sprintf(item.saved_text, "%s", SLIDER_LEFT);
-				for (int j = 0; j < (item.max_value - item.min_value + 1); j++)
-				{
-					sprintf(item.saved_text, "%s%s", item.saved_text, SLIDER_MIDDLE);
-				}
-				sprintf(item.saved_text, "%s%s", item.saved_text, SLIDER_RIGHT);
-				gr_get_string_size(item.saved_text, &w1, &h1, &aw1);
-				string_width += w1 + aw;
-			}
-
-			if (item.type == NM_TYPE_CHECK)
-			{
-				int w1, h1, aw1;
-				gr_get_string_size(NORMAL_CHECK_BOX, &w1, &h1, &aw1);
-				item.right_offset = w1;
-				gr_get_string_size(CHECKED_CHECK_BOX, &w1, &h1, &aw1);
-				if (w1 > item.right_offset)
-					item.right_offset = w1;
-			}
-
-			if (item.type == NM_TYPE_RADIO)
-			{
-				int w1, h1, aw1;
-				gr_get_string_size(NORMAL_RADIO_BOX, &w1, &h1, &aw1);
-				item.right_offset = w1;
-				gr_get_string_size(CHECKED_RADIO_BOX, &w1, &h1, &aw1);
-				if (w1 > item.right_offset)
-					item.right_offset = w1;
-			}
-
-			if (item.type == NM_TYPE_NUMBER)
-			{
-				int w1, h1, aw1;
-				char test_text[20];
-				sprintf(test_text, "%d", item.max_value);
-				gr_get_string_size(test_text, &w1, &h1, &aw1);
-				item.right_offset = w1;
-				sprintf(test_text, "%d", item.min_value);
-				gr_get_string_size(test_text, &w1, &h1, &aw1);
-				if (w1 > item.right_offset)
-					item.right_offset = w1;
-			}
-
-			if (item.type == NM_TYPE_INPUT)
-			{
-				Assert(strlen(item.text) < NM_MAX_TEXT_LEN);
-				strcpy(item.saved_text, item.text);
-				string_width = item.text_len * grd_curcanv->cv_font->ft_w + item.text_len;
-				if (string_width > MAX_TEXT_WIDTH)
-					string_width = MAX_TEXT_WIDTH;
-				item.value = -1;
-			}
-
-			if (item.type == NM_TYPE_INPUT_MENU)
-			{
-				Assert(strlen(item.text) < NM_MAX_TEXT_LEN);
-				strcpy(item.saved_text, item.text);
-				string_width = item.text_len * grd_curcanv->cv_font->ft_w + item.text_len;
-				item.value = -1;
-				item.group = 0;
-			}
-
-			item.w = string_width;
-			item.h = string_height;
-
-			if (string_width > w)
-				w = string_width;		// Save maximum width
-			if (average_width > aw)
-				aw = average_width;
-			h += string_height + 1;		// Find the height of all strings
-		}
-
-		int right_offset = 0;
-
-		if (width > -1)
-			w = width;
-
-		if (height > -1)
-			h = height;
-
-		for (newmenu_item& item : items)
-		{
-			item.w = w;
-			if (item.right_offset > right_offset)
-				right_offset = item.right_offset;
-		}
-		if (right_offset > 0)
-			right_offset += 3;
-
-		//mprintf( 0, "Right offset = %d\n", right_offset );
-
-		//gr_get_string_size("",&string_width,&string_height,&average_width );
-
-		w += right_offset;
-
-		int twidth = 0;
-		if (tw > w)
-		{
-			twidth = (tw - w) / 2;
-			w = tw;
-		}
-
-		// Find min point of menu border
-		w += 30;
-		h += 30;
-
-		if (w > 320) w = 320;
-		if (h > 200) h = 200;
-
-		x = (grd_curcanv->cv_bitmap.bm_w - w) / 2;
-		y = (grd_curcanv->cv_bitmap.bm_h - h) / 2;
-
-		if (x < 0) x = 0;
-		if (y < 0) y = 0;
-		if (filename != NULL)
-		{
-			nm_draw_background1(filename);
-			bg_filename = std::string(filename);
-		}
-
-		// Save the background of the display
-		bg.menu_canvas = gr_create_sub_canvas(nm_canvas, x, y, w, h);
-		gr_set_current_canvas(bg.menu_canvas);
-
-		if (filename == NULL)
-		{
-			// Save the background under the menu...
-			bg.saved = gr_create_bitmap(w, h);
-			Assert(bg.saved != NULL);
-			gr_bm_bitblt(w, h, 0, 0, 0, 0, &grd_curcanv->cv_bitmap, bg.saved);
-			gr_set_current_canvas(nm_canvas);
-			nm_draw_background(x, y, x + w - 1, y + h - 1);
-			bg.background = gr_create_sub_bitmap(&nm_background, 0, 0, w, h);
-			gr_set_current_canvas(bg.menu_canvas);
-		}
-		else
-		{
-			bg.saved = NULL;
-			bg.background = gr_create_bitmap(w, h);
-			Assert(bg.background != NULL);
-			gr_bm_bitblt(w, h, 0, 0, 0, 0, &grd_curcanv->cv_bitmap, bg.background);
-		}
-
-		int ty = 15;
-
-		if (title.size() > 0)
-		{
-			grd_curcanv->cv_font = TITLE_FONT;
-			gr_set_fontcolor(GR_GETCOLOR(31, 31, 31), -1);
-			gr_get_string_size(title.c_str(), &string_width, &string_height, &average_width);
-			tw = string_width;
-			th = string_height;
-			gr_printf(0x8000, ty, title.c_str());
-			ty += th;
-		}
-
-		if (subtitle.size() > 0)
-		{
-			grd_curcanv->cv_font = SUBTITLE_FONT;
-			gr_set_fontcolor(GR_GETCOLOR(21, 21, 21), -1);
-			gr_get_string_size(subtitle.c_str(), &string_width, &string_height, &average_width);
-			tw = string_width;
-			th = string_height;
-			gr_printf(0x8000, ty, subtitle.c_str());
-			ty += th;
-		}
-
-		if (tiny_mode)
-			grd_curcanv->cv_font = SMALL_FONT;
-		else
-			grd_curcanv->cv_font = NORMAL_FONT;
-
-		// Update all item's x & y values.
-		for (newmenu_item& item : items)
-		{
-			item.x = 15 + twidth + right_offset;
-			item.y += 15;
-			if (item.type == NM_TYPE_RADIO)
-			{
-				int fm = -1;	// find first marked one
-				for (int j = 0; j < items.size(); j++)
-				{
-					if (items[j].type == NM_TYPE_RADIO && items[j].group == item.group)
-					{
-						if (fm == -1 && items[j].value)
-							fm = j;
-						items[j].value = 0;
-					}
-				}
-				if (fm >= 0)
-					items[fm].value = 1;
-				else
-					item.value = 1;
-			}
-		}
-
-		if (citem == -1)
-		{
-			choice = -1;
-		}
-		else
-		{
-			if (citem < 0) citem = 0;
-			if (citem > items.size() - 1) citem = items.size() - 1;
-			choice = citem;
-
-			while (items[choice].type == NM_TYPE_TEXT)
-			{
-				choice++;
-				if (choice >= items.size())
-				{
-					choice = 0;
-				}
-				if (choice == citem)
-				{
-					choice = 0;
-					all_text = true;
-					break;
-				}
-			}
-		}
-
-		gr_set_current_canvas(save_canvas);
-	}
-
-	void frame() override
-	{
-		bool done = false;
-		int k = key_inkey();
-
-		gr_set_current_canvas(bg.menu_canvas);
-
-		if (newmenu_callback)
-			(*newmenu_callback)(items.size(), items.data(), &k, choice);
-
-		if (k < -1)
-		{
-			choice = k;
-			k = -1;
-			done = true;
-		}
-
-		old_choice = choice;
-
-		switch (k)
-		{
-		case KEY_TAB + KEY_SHIFTED:
-		case KEY_UP:
-		case KEY_PAD8:
-			if (all_text) break;
-			do
-			{
-				choice--;
-				if (choice >= items.size()) choice = 0;
-				if (choice < 0) choice = items.size() - 1;
-			} while (items[choice].type == NM_TYPE_TEXT);
-			if ((items[choice].type == NM_TYPE_INPUT) && (choice != old_choice))
-				items[choice].value = -1;
-			if ((old_choice > -1) && (items[old_choice].type == NM_TYPE_INPUT_MENU) && (old_choice != choice))
-			{
-				items[old_choice].group = 0;
-				strcpy(items[old_choice].text, items[old_choice].saved_text);
-				items[old_choice].value = -1;
-			}
-			if (old_choice > -1)
-				items[old_choice].redraw = 1;
-			items[choice].redraw = 1;
-			break;
-		case KEY_TAB:
-		case KEY_DOWN:
-		case KEY_PAD2:
-			if (all_text) break;
-			do
-			{
-				choice++;
-				if (choice < 0) choice = items.size() - 1;
-				if (choice >= items.size()) choice = 0;
-			} while (items[choice].type == NM_TYPE_TEXT);
-			if ((items[choice].type == NM_TYPE_INPUT) && (choice != old_choice))
-				items[choice].value = -1;
-			if ((old_choice > -1) && (items[old_choice].type == NM_TYPE_INPUT_MENU) && (old_choice != choice))
-			{
-				items[old_choice].group = 0;
-				strcpy(items[old_choice].text, items[old_choice].saved_text);
-				items[old_choice].value = -1;
-			}
-			if (old_choice > -1)
-				items[old_choice].redraw = 1;
-			items[choice].redraw = 1;
-			break;
-		case KEY_SPACEBAR:
-			if (choice > -1)
-			{
-				switch (items[choice].type)
-				{
-				case NM_TYPE_MENU:
-				case NM_TYPE_INPUT:
-				case NM_TYPE_INPUT_MENU:
-					break;
-				case NM_TYPE_CHECK:
-					if (items[choice].value)
-						items[choice].value = 0;
-					else
-						items[choice].value = 1;
-					items[choice].redraw = 1;
-					break;
-				case NM_TYPE_RADIO:
-					for (int i = 0; i < items.size(); i++)
-					{
-						if ((i != choice) && (items[i].type == NM_TYPE_RADIO) && (items[i].group == items[choice].group) && (items[i].value))
-						{
-							items[i].value = 0;
-							items[i].redraw = 1;
-						}
-					}
-					items[choice].value = 1;
-					items[choice].redraw = 1;
-					break;
-				}
-			}
-			break;
-
-		case KEY_ENTER:
-		case KEY_PADENTER:
-			if ((choice > -1) && (items[choice].type == NM_TYPE_INPUT_MENU) && (items[choice].group == 0))
-			{
-				items[choice].group = 1;
-				items[choice].redraw = 1;
-				if (!_strnicmp(items[choice].saved_text, TXT_EMPTY, strlen(TXT_EMPTY)))
-				{
-					items[choice].text[0] = 0;
-					items[choice].value = -1;
-				}
-				else
-				{
-					strip_end_whitespace(items[choice].text);
-				}
-			}
-			else
-				done = true;
-			break;
-
-		case KEY_ESC:
-			if ((choice > -1) && (items[choice].type == NM_TYPE_INPUT_MENU) && (items[choice].group == 1))
-			{
-				items[choice].group = 0;
-				strcpy(items[choice].text, items[choice].saved_text);
-				items[choice].redraw = 1;
-				items[choice].value = -1;
-			}
-			else
-			{
-				done = true;
-				choice = -1;
-			}
-			break;
-
-		case KEY_PRINT_SCREEN: 		save_screen_shot(0); break;
-
-#ifndef NDEBUG
-		case KEY_BACKSP:
-			if ((choice > -1) && (items[choice].type != NM_TYPE_INPUT) && (items[choice].type != NM_TYPE_INPUT_MENU))
-				Int3();
-			break;
-#endif
-
-		}
-
-		if (choice > -1)
-		{
-			int ascii;
-
-			if (((items[choice].type == NM_TYPE_INPUT) || ((items[choice].type == NM_TYPE_INPUT_MENU) && (items[choice].group == 1))) && (old_choice == choice))
-			{
-				if (k == KEY_LEFT || k == KEY_BACKSP || k == KEY_PAD4)
-				{
-					if (items[choice].value == -1) items[choice].value = strlen(items[choice].text);
-					if (items[choice].value > 0)
-						items[choice].value--;
-					items[choice].text[items[choice].value] = 0;
-					items[choice].redraw = 1;
-				}
-				else
-				{
-					ascii = key_to_ascii(k);
-					if ((ascii < 255) && (items[choice].value < items[choice].text_len))
-					{
-						int allowed;
-
-						if (items[choice].value == -1)
-						{
-							items[choice].value = 0;
-						}
-
-						allowed = char_allowed(ascii);
-
-						if (!allowed && ascii == ' ' && char_allowed('_'))
-						{
-							ascii = '_';
-							allowed = 1;
-						}
-
-						if (allowed)
-						{
-							items[choice].text[items[choice].value++] = ascii;
-							items[choice].text[items[choice].value] = 0;
-							items[choice].redraw = 1;
-						}
-					}
-				}
-			}
-			else if ((items[choice].type != NM_TYPE_INPUT) && (items[choice].type != NM_TYPE_INPUT_MENU))
-			{
-				ascii = key_to_ascii(k);
-				if (ascii < 255)
-				{
-					int choice1 = choice;
-					ascii = toupper(ascii);
-					do
-					{
-						int i, ch;
-						choice1++;
-						if (choice1 >= items.size()) choice1 = 0;
-						for (i = 0; (ch = items[choice1].text[i]) != 0 && ch == ' '; i++);
-						if (((items[choice1].type == NM_TYPE_MENU) ||
-							(items[choice1].type == NM_TYPE_CHECK) ||
-							(items[choice1].type == NM_TYPE_RADIO) ||
-							(items[choice1].type == NM_TYPE_NUMBER) ||
-							(items[choice1].type == NM_TYPE_SLIDER))
-							&& (ascii == toupper(ch)))
-						{
-							k = 0;
-							choice = choice1;
-							if (old_choice > -1)
-								items[old_choice].redraw = 1;
-							items[choice].redraw = 1;
-						}
-					} while (choice1 != choice);
-				}
-			}
-
-			if ((items[choice].type == NM_TYPE_NUMBER) || (items[choice].type == NM_TYPE_SLIDER))
-			{
-				int ov = items[choice].value;
-				switch (k)
-				{
-				case KEY_PAD4:
-				case KEY_LEFT:
-				case KEY_MINUS:
-				case KEY_MINUS + KEY_SHIFTED:
-				case KEY_PADMINUS:
-					items[choice].value -= 1;
-					break;
-				case KEY_RIGHT:
-				case KEY_PAD6:
-				case KEY_EQUAL:
-				case KEY_EQUAL + KEY_SHIFTED:
-				case KEY_PADPLUS:
-					items[choice].value++;
-					break;
-				case KEY_PAGEUP:
-				case KEY_PAD9:
-				case KEY_SPACEBAR:
-					items[choice].value += 10;
-					break;
-				case KEY_PAGEDOWN:
-				case KEY_BACKSP:
-				case KEY_PAD3:
-					items[choice].value -= 10;
-					break;
-				}
-				if (ov != items[choice].value)
-					items[choice].redraw = 1;
-			}
-
-		}
-
-		gr_set_current_canvas(bg.menu_canvas);
-		// Redraw everything...
-		for (int i = 0; i < items.size(); i++)
-		{
-			if (items[i].redraw)
-			{
-				draw_item(&bg, &items[i], (i == choice && !all_text), tiny);
-				items[i].redraw = 0;
-			}
-			else if (i == choice && (items[i].type == NM_TYPE_INPUT || (items[i].type == NM_TYPE_INPUT_MENU && items[i].group)))
-				update_cursor(&items[i]);
-		}
-
-		if (done) //A choice was made
-		{
-			bool keep_open = choice_callback(choice);
-			if (keep_open)
-			{
-				if (choice == -1) //hit escape, but remaining open?
-					choice = old_choice; 
-			}
-			else
-			{
-				close(); 
-			}
-		}
-	}
-
-	void cleanup() override
-	{
-		// Restore everything...
-		gr_set_current_canvas(bg.menu_canvas);
-		if (bg_filename.size() == 0)
-		{
-			// Save the background under the menu...
-			gr_bitmap(0, 0, bg.saved);
-			gr_free_bitmap(bg.saved);
-			free(bg.background);
-		}
-		else
-		{
-			gr_bitmap(0, 0, bg.background);
-			gr_free_bitmap(bg.background);
-		}
-
-		gr_free_sub_canvas(bg.menu_canvas);
-	}
-};
-
-void newmenu_open(const char* title, const char* subtitle, std::vector<newmenu_item>& items, void (*subfunction)(int nitems, newmenu_item* items, int* last_key, int citem), bool (*choicefunc)(int choice), int citem, const char* filename)
+void newmenu_open(const char* title, const char* subtitle, int nitems, newmenu_item* items, void (*subfunction)(int nitems, newmenu_item* items, int* last_key, int citem), bool (*choicefunc)(int choice, newmenu_item* item), int citem, const char* filename, int width, int height, bool tiny_mode)
 {
-	nm_open_windows.push(std::make_unique<nm_menu>(items, title, subtitle, subfunction, choicefunc, citem, filename, -1, -1, false));
+	std::vector<newmenu_item> itemclone;
+	for (int i = 0; i < nitems; i++)
+		itemclone.push_back(items[i]);
+
+	nm_open_windows.push(std::make_unique<nm_menu>(itemclone, title, subtitle, subfunction, choicefunc, citem, filename, width, height, tiny_mode));
 }
 
 //Simple callback for purely informative message boxes
-bool newmenu_messagebox_informative_callback(int choice)
+bool newmenu_messagebox_informative_callback(int choice, newmenu_item* item)
 {
 	return false; //never remain open. 
 }
 
-void nm_open_messagebox(const char* title, bool (*callback)(int choice), int nchoices, ...)
+void nm_open_messagebox(const char* title, bool (*callback)(int choice, newmenu_item* item), int nchoices, ...)
 {
 	va_list args;
 	char nm_text[MESSAGEBOX_TEXT_SIZE];
@@ -1149,470 +446,6 @@ void nm_open_messagebox(const char* title, bool (*callback)(int choice), int nch
 }
 
 
-//-----------------------------------------------------------------------------
-// nm_filelist
-//
-// Implements a list box that lists files. 
-//-----------------------------------------------------------------------------
-
-struct newmenu_fileentry
-{
-	std::string filename;
-	std::string displayname;
-	bool truncated;
-
-	newmenu_fileentry()
-	{
-		truncated = false;
-	}
-
-	newmenu_fileentry(std::string_view newfilename, bool noext)
-	{
-		filename = newfilename;
-		truncated = noext;
-		if (noext)
-		{
-			std::string::size_type pos = filename.find_last_of(".");
-			if (pos == std::string::npos)
-				displayname = filename;
-			else
-				displayname = std::string(filename, 0, pos);
-		}
-		else
-		{
-			displayname = filename;
-		}
-	}
-
-	//Helper function, truncated entries can have a $ as the first character, this needs to be stored but not displayed
-	const char* get_display_name()
-	{
-		if (!truncated || (displayname.size() > 0 && displayname[0] != '$'))
-			return displayname.c_str();
-
-		return displayname.c_str() + 1;
-	}
-};
-
-void newmenu_file_sort(std::span<newmenu_fileentry> entries)
-{
-	int n = entries.size();
-	int incr = n / 2;
-	newmenu_fileentry t;
-
-	while (incr > 0)
-	{
-		for (int i = incr; i < n; i++)
-		{
-			int j = i - incr;
-			while (j >= 0)
-			{
-				if (entries[j].filename.compare(entries[j + incr].filename) > 0)
-				{
-					t = std::move(entries[j]);
-					entries[j] = std::move(entries[j + incr]);
-					entries[j + incr] = std::move(t);
-					j -= incr;
-				}
-				else
-					break;
-			}
-		}
-		incr /= 2;
-	}
-}
-
-void delete_player_saved_games(const char* name)
-{
-	int i;
-	char filename_full_path[CHOCOLATE_MAX_FILE_PATH_SIZE];
-	char filename[16];
-
-	for (i = 0; i < 10; i++)
-	{
-		snprintf(filename, 16, "%s.sg%d", name, i);
-		get_game_full_file_path(filename_full_path, filename, CHOCOLATE_SAVE_DIR);
-		_unlink(filename_full_path);
-		//sprintf(filename, "%s.sg%d", name, i);
-		//_unlink(filename);
-	}
-}
-
-//This isn't particularly elegant, but there will only be one nm_filelist active with pending deletes
-static bool filelist_do_delete = false;
-
-bool nm_delete_callback(int num)
-{
-	if (num == 0)
-		filelist_do_delete = true;
-
-	return false; //Never keep the window open
-}
-
-class nm_filelist : public nm_window
-{
-	std::vector<newmenu_fileentry> files;
-	int NumFiles;
-	int NumFiles_displayed = 8;
-	int first_item = -1;
-	bool player_mode = false;
-	bool demo_mode = false;
-	int demos_deleted = 0;
-	int initialized = 0;
-	int exit_value = 0;
-	int w_x, w_y, w_w, w_h;
-	int citem;
-	bkg bg;
-	bool allow_abort_flag;
-
-	std::string title;
-	char localized_filespec[CHOCOLATE_MAX_FILE_PATH_SIZE];
-
-	bool pending_delete;
-	int pending_delete_num;
-
-	void (*file_selected)(std::string& filename, int num);
-
-	void create_list()
-	{
-		files.clear();
-		NumFiles = 0;
-
-		if (player_mode)
-		{
-			files.emplace_back(TXT_CREATE_NEW, false);
-			NumFiles++;
-		}
-
-		FILEFINDSTRUCT find;
-		if (!FileFindFirstLFNTemp(localized_filespec, &find))
-			//if (!FileFindFirstLFNTemp(filespec, &find)) 
-		{
-			do
-			{
-				if (!player_mode || strlen(find.name) <= 13) //12345678.0123
-				{
-					files.emplace_back(find.name, player_mode);
-					NumFiles++;
-				}
-			} while (!FileFindNext(&find));
-			FileFindClose();
-		}
-
-		if ((NumFiles < 1) && demos_deleted)
-		{
-			exit_value = 0;
-			//Just close up if this happened
-			close();
-			//goto ExitFileMenu;
-		}
-		if ((NumFiles < 1) && demo_mode)
-		{
-			Int3();
-			nm_messagebox(NULL, 1, TXT_OK, "%s %s\n%s", TXT_NO_DEMO_FILES, TXT_USE_F5, TXT_TO_CREATE_ONE);
-			exit_value = 0;
-			//goto ExitFileMenu;
-		}
-
-		if ((NumFiles < 2) && player_mode)
-		{
-			citem = 0;
-			file_selected(files[0].filename, citem);
-			//goto ExitFileMenuEarly;
-		}
-
-		if (NumFiles < 1)
-		{
-			nm_messagebox(NULL, 1, "Ok", "%s\n '%s' %s", TXT_NO_FILES_MATCHING, localized_filespec, TXT_WERE_FOUND);
-		}
-
-		if (!player_mode)
-		{
-			newmenu_file_sort(files);
-		}
-		else
-		{
-			newmenu_file_sort(std::span<newmenu_fileentry>(files.begin() + 1, files.end()));
-
-			for (int i = 0; i < files.size(); i++)
-			{
-				if (!_strfcmp(Players[Player_num].callsign, files[i].displayname.c_str()))
-					citem = i;
-			}
-		}
-	}
-
-public:
-	nm_filelist(const char* newtitle, const char* filespec, void (*select_callback)(std::string& filename, int num), bool allow_abort)
-	{
-		title = std::string(newtitle);
-		file_selected = select_callback;
-		citem = 0;
-		allow_abort_flag = allow_abort;
-
-		//Get localized filepath
-		const char* wildcard_pos;
-		wildcard_pos = strrchr(filespec, '*');
-		if (wildcard_pos != NULL)
-		{
-			if (!_strfcmp(wildcard_pos, "*.nplt"))
-				player_mode = true;
-			else if (!_strfcmp(wildcard_pos, "*.dem"))
-				demo_mode = true;
-		}
-		strncpy(localized_filespec, filespec, CHOCOLATE_MAX_FILE_PATH_SIZE);
-		localized_filespec[CHOCOLATE_MAX_FILE_PATH_SIZE - 1] = '\0';
-
-		create_list();
-
-		gr_set_current_canvas(nm_canvas);
-
-		w_w = 230 - 90 + 1 + 30;
-		w_h = 170 - 30 + 1 + 30;
-
-		if (w_w > 320) w_w = 320;
-		if (w_h > 200) w_h = 200;
-
-		w_x = (grd_curcanv->cv_bitmap.bm_w - w_w) / 2;
-		w_y = (grd_curcanv->cv_bitmap.bm_h - w_h) / 2;
-
-		if (w_x < 0) w_x = 0;
-		if (w_y < 0) w_y = 0;
-
-		// Save the background under the menu...
-		//TODO: Since the file list doesn't use a sub canvas, this creates a large region than expected. 
-		bg.saved = gr_create_bitmap(nm_canvas->cv_bitmap.bm_w, nm_canvas->cv_bitmap.bm_h);
-		Assert(bg.saved != NULL);
-		gr_set_current_canvas(nm_canvas);
-		gr_bm_bitblt(nm_canvas->cv_bitmap.bm_w, nm_canvas->cv_bitmap.bm_h, 0, 0, 0, 0, &grd_curcanv->cv_bitmap, bg.saved);
-
-		nm_draw_background(w_x, w_y, w_x + w_w - 1, w_y + w_h - 1);
-
-		grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_3];
-		gr_string(0x8000, w_y + 10, title.c_str());
-
-		pending_delete = false;
-		pending_delete_num = 0;
-	}
-
-	void frame() override
-	{
-		bool done = false;
-		int ocitem = citem;
-		int ofirst_item = first_item;
-		int key = key_inkey();
-		bool redraw = false;
-
-		gr_set_current_canvas(nm_canvas);
-
-		if (pending_delete)
-		{
-			if (filelist_do_delete)
-			{
-				char file_full_path[CHOCOLATE_MAX_FILE_PATH_SIZE];
-
-				if (player_mode)
-					get_full_file_path(file_full_path, files[citem].filename.c_str(), CHOCOLATE_PILOT_DIR);
-				else if (demo_mode)
-					get_game_full_file_path(file_full_path, files[citem].filename.c_str(), CHOCOLATE_DEMOS_DIR);
-				int ret = _unlink(file_full_path);
-				//ret = _unlink(files[citem].filename.c_str());
-
-				if ((!ret) && player_mode)
-				{
-					delete_player_saved_games(files[citem].filename.c_str());
-				}
-
-				if (ret)
-				{
-					if (player_mode)
-						nm_open_messagebox(nullptr, nullptr, 1, TXT_OK, "%s %s %s", TXT_COULDNT, TXT_DELETE_PILOT, files[citem].get_display_name());
-					else if (demo_mode)
-						nm_open_messagebox(nullptr, nullptr, 1, TXT_OK, "%s %s %s", TXT_COULDNT, TXT_DELETE_DEMO, files[citem].get_display_name());
-				}
-				else if (demo_mode)
-					demos_deleted = 1;
-				first_item = -1;
-				//Regenerate the list
-				create_list();
-				filelist_do_delete = false;
-				redraw = true;
-			}
-			pending_delete = false;
-		}
-
-		switch (key)
-		{
-		case KEY_PRINT_SCREEN: 		save_screen_shot(0); break;
-		case KEY_CTRLED + KEY_D:
-			if (((player_mode) && (citem > 0)) || ((demo_mode) && (citem >= 0)))
-			{
-				if (player_mode)
-				{
-					nm_open_messagebox(NULL, nm_delete_callback, 2, TXT_YES, TXT_NO, "%s %s?", TXT_DELETE_PILOT, files[citem].get_display_name());
-					pending_delete = true;
-					pending_delete_num = citem;
-				}
-				else if (demo_mode)
-				{
-					nm_open_messagebox(NULL, nm_delete_callback, 2, TXT_YES, TXT_NO, "%s %s?", TXT_DELETE_DEMO, files[citem].get_display_name());
-					pending_delete = true;
-					pending_delete_num = citem;
-				}
-			}
-			break;
-		case KEY_HOME:
-		case KEY_PAD7:
-			citem = 0;
-			break;
-		case KEY_END:
-		case KEY_PAD1:
-			citem = NumFiles - 1;
-			break;
-		case KEY_UP:
-		case KEY_PAD8:
-			citem--;
-			break;
-		case KEY_DOWN:
-		case KEY_PAD2:
-			citem++;
-			break;
-		case KEY_PAGEDOWN:
-		case KEY_PAD3:
-			citem += NumFiles_displayed;
-			break;
-		case KEY_PAGEUP:
-		case KEY_PAD9:
-			citem -= NumFiles_displayed;
-			break;
-		case KEY_ESC:
-			if (allow_abort_flag)
-			{
-				citem = -1;
-				done = true;
-			}
-			break;
-		case KEY_ENTER:
-		case KEY_PADENTER:
-			done = true;
-			break;
-		default:
-		{
-			int ascii = key_to_ascii(key);
-			if (ascii < 255)
-			{
-				int cc, cc1;
-				cc = cc1 = citem + 1;
-				if (cc1 < 0)  cc1 = 0;
-				if (cc1 >= NumFiles)  cc1 = 0;
-				while (1) {
-					if (cc < 0) cc = 0;
-					if (cc >= NumFiles) cc = 0;
-					if (citem == cc) break;
-
-					if (toupper(files[cc].displayname[0]) == toupper(ascii))
-					{
-						citem = cc;
-						break;
-					}
-					cc++;
-				}
-			}
-		}
-		}
-
-		if (done)
-		{
-			file_selected(files[citem].filename, citem);
-			close();
-		}
-		else
-		{
-			if (citem < 0)
-				citem = 0;
-
-			if (citem >= NumFiles)
-				citem = NumFiles - 1;
-
-			if (citem < first_item)
-				first_item = citem;
-
-			if (citem >= (first_item + NumFiles_displayed))
-				first_item = citem - NumFiles_displayed + 1;
-
-			if (NumFiles <= NumFiles_displayed)
-				first_item = 0;
-
-			if (first_item > NumFiles - NumFiles_displayed)
-				first_item = NumFiles - NumFiles_displayed;
-			if (first_item < 0) first_item = 0;
-
-			if (ofirst_item != first_item || redraw)
-			{
-				gr_setcolor(BM_XRGB(0, 0, 0));
-				for (int i = first_item; i < first_item + NumFiles_displayed; i++)
-				{
-					int w, h, aw, y;
-					y = (i - first_item) * 12 + w_y + 45;
-					if (i >= NumFiles)
-					{
-						gr_setcolor(BM_XRGB(0, 0, 0));
-						gr_rect(100, y - 1, 220, y + 11);
-					}
-					else
-					{
-						if (i == citem)
-							grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_2];
-						else
-							grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_1];
-						gr_get_string_size(files[i].displayname.c_str(), &w, &h, &aw);
-						gr_rect(100, y - 1, 220, y + 11);
-						//gr_string(105, y, (&filenames[i * 14]) + ((player_mode && filenames[i * 14] == '$') ? 1 : 0));
-						gr_string(105, y, files[i].get_display_name());
-					}
-				}
-			}
-			else if (citem != ocitem)
-			{
-				int w, h, aw, y;
-
-				int i = ocitem;
-				if ((i >= 0) && (i < NumFiles))
-				{
-					y = (i - first_item) * 12 + w_y + 45;
-					if (i == citem)
-						grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_2];
-					else
-						grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_1];
-					gr_get_string_size(files[i].displayname.c_str(), &w, &h, &aw);
-					gr_rect(100, y - 1, 220, y + 11);
-					//gr_string(105, y, (&filenames[i * 14]) + ((player_mode && filenames[i * 14] == '$') ? 1 : 0));
-					gr_string(105, y, files[i].get_display_name());
-				}
-				i = citem;
-				if ((i >= 0) && (i < NumFiles))
-				{
-					y = (i - first_item) * 12 + w_y + 45;
-					if (i == citem)
-						grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_2];
-					else
-						grd_curcanv->cv_font = Gamefonts[GFONT_MEDIUM_1];
-					gr_get_string_size(files[i].displayname.c_str(), &w, &h, &aw);
-					gr_rect(100, y - 1, 220, y + 11);
-					gr_string(105, y, files[i].get_display_name());
-				}
-			}
-}
-	}
-
-	void cleanup() override
-	{
-		// Restore everything...
-		gr_bitmap(0, 0, bg.saved);
-		gr_free_bitmap(bg.saved);
-	}
-};
-
 void newmenu_open_filepicker(const char* title, const char* filespec, int allow_abort_flag, void (*callback)(std::string& str, int num))
 {
 	nm_open_windows.push(std::make_unique<nm_filelist>(title, filespec, callback, allow_abort_flag));
@@ -1630,6 +463,10 @@ void newmenu_frame()
 		}
 		else
 		{
+			//Drawing should be delayed, otherwise opening a window while closing the current won't work right
+			if (top->must_draw())
+				top->draw();
+
 			top->frame(); //Run the window frame. 
 			did_frame = true;
 		}
@@ -1639,6 +476,11 @@ void newmenu_frame()
 void newmenu_present()
 {
 	plat_present_canvas_masked(*nm_canvas, 3.f / 4.f);
+}
+
+bool newmenu_empty()
+{
+	return nm_open_windows.empty();
 }
 
 int newmenu_do3(const char* title, const char* subtitle, int nitems, newmenu_item* item, void (*subfunction)(int nitems, newmenu_item* items, int* last_key, int citem), int citem, const char* filename, int width, int height, bool tiny_mode)
@@ -1701,7 +543,7 @@ int nm_messagebox(const char* title, int nchoices, ...)
 
 #define MAX_FILES 100
 
-int MakeNewPlayerFile(int allow_abort);
+void MakeNewPlayerFile(int allow_abort);
 
 int newmenu_get_filename(const char* title, const char* filespec, char* filename, int allow_abort_flag)
 {
